ai-server/requirements.txt
numpy
scikit-learn
joblib
opencv-python
mediapipe
flask
flask-cors
pymongo
python-dotenv
